requirements.txt
biopython==1.84
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
tqdm==4.66.1
# torch==2.4.1
hydra-core==1.3.2
transformers==4.29
huggingface-hub==0.23.4
wandb==0.13.5